fix:统一方法的参数位置和参数类型

+++ moduleImageLoader/src/main/java/com/renxing/moduleImageLoader/loaderStrategy/glide/transformation/BorderRoundTransform.java
@@ -28,7 +28,7 @@ public class BorderRoundTransform extends BitmapTransformation {
     private float mRadius;
 
 
-    public BorderRoundTransform(float borderWidth, int borderColor, int cornerWidth) {
+    public BorderRoundTransform(float borderWidth, int borderColor, float cornerWidth) {
         this.borderWidth = borderWidth;
         this.borderColor = borderColor;
         mRadius = Resources.getSystem().getDisplayMetrics().density * cornerWidth;
